Reorganized tools and doc-building logic.
