--- supriya/tools/dawtools/__init__.py
@@ -1,12 +1,0 @@
-# -*- encoding: utf-8 -*-
-
-"""
-High-level tools for synths, effects, monitoring and mixing.
-"""
-
-from abjad.tools import systemtools
-
-systemtools.ImportManager.import_structured_package(
-    __path__[0],
-    globals(),
-    )
