--- supriya/tools/documentationtools/__init__.py
@@ -1,12 +1,0 @@
-# -*- encoding: utf-8 -*-
-
-"""
-Tools for creating Supriya's documentation.
-"""
-
-from abjad.tools import systemtools
-
-systemtools.ImportManager.import_structured_package(
-    __path__[0],
-    globals(),
-    )
